refactor(bayn): split candidate development domain core
